[svn-r12314] Purpose:     Code cleanup
Description:
    Big clean up on high-level library code, including:
        - Removing include of HDF5 private header file(!)
        - Reduce number of compiler warnings
        - Untangle header files, etc.

Platforms tested:
    FreeBSD 4.11 (sleipnir) w/C++
    Linux 2.4/64 (mir) w/C++ & FORTRAN

+++ hl/tools/gif2h5/gifread.c
@@ -65,7 +65,7 @@ ReadGifHeader(GIFHEAD *GifHead, /* Pointer to GIF header structure  */
         GifHead->HeaderDump[i] = *(*MemGif2)++;
     }
 
-    if (strncmp((const char *)GifHead->HeaderDump , "GIF" , 3)) {
+    if (strncmp((const char *)GifHead->HeaderDump , "GIF" , (size_t)3)) {
         printf("The file does not appear to be a valid GIF file.\n");
         exit(-1);
     }
