+++ hl/tools/gif2h5/hdfgifwr.c
@@ -515,7 +515,7 @@ flush_char(void)
 {
     if (a_count > 0) {
         fputc( a_count, g_outfile );
-        fwrite( accum, 1, (size_t)a_count, g_outfile);
+        fwrite( accum, (size_t)1, (size_t)a_count, g_outfile);
         a_count = 0;
     }
 }
